+++ apps/e2e/tests/backend/endpoints/api/v1/users.test.ts
@@ -1,7 +1,7 @@
 import { generateSecureRandomString } from "@stackframe/stack-shared/dist/utils/crypto";
 import { describe } from "vitest";
-import { createMailbox, it } from "../../../../helpers";
-import { Auth, InternalProjectKeys, Project, backendContext, niceBackendFetch } from "../../../backend-helpers";
+import { it } from "../../../../helpers";
+import { Auth, InternalProjectKeys, Project, backendContext, bumpEmailAddress, createMailbox, niceBackendFetch } from "../../../backend-helpers";
 
 describe("without project access", () => {
   backendContext.set({
@@ -31,6 +31,7 @@ describe("without project access", () => {
   });
 
   it("should not be able to list users", async ({ expect }) => {
+    await Project.createAndSwitch();
     const response = await niceBackendFetch("/api/v1/users");
     expect(response).toMatchInlineSnapshot(`
       NiceResponse {
@@ -86,7 +87,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -118,7 +119,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -195,7 +196,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -226,7 +227,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -367,7 +368,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -398,7 +399,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -450,6 +451,7 @@ describe("with client access", () => {
   });
 
   it("should not be able to list users", async ({ expect }) => {
+    await Project.createAndSwitch();
     const response = await niceBackendFetch("/api/v1/users", {
       accessType: "client",
     });
@@ -527,7 +529,7 @@ describe("with client access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
           "requires_totp_mfa": false,
@@ -636,7 +638,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -674,7 +676,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -728,7 +730,7 @@ describe("with server access", () => {
               "oauth_providers": [],
               "otp_auth_enabled": true,
               "passkey_auth_enabled": false,
-              "primary_email": "<stripped UUID>@stack-generated.example.com",
+              "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
               "primary_email_auth_enabled": true,
               "primary_email_verified": true,
               "profile_image_url": null,
@@ -746,10 +748,10 @@ describe("with server access", () => {
     `);
   });
 
-  it("list next cursor", async ({ expect }) => {
+  it("lists users with pagination", async ({ expect }) => {
     await Project.createAndSwitch({ config: { magic_link_enabled: true } });
     for (let i = 0; i < 5; i++) {
-      backendContext.set({ mailbox: createMailbox() });
+      await bumpEmailAddress();
       await Auth.Otp.signIn();
     }
     const allResponse = await niceBackendFetch("/api/v1/users", {
@@ -799,7 +801,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -875,7 +877,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -956,7 +958,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1008,7 +1010,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1114,7 +1116,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1140,7 +1142,7 @@ describe("with server access", () => {
         "status": 400,
         "body": {
           "code": "USER_EMAIL_ALREADY_EXISTS",
-          "error": "User already exists.",
+          "error": "User email already exists.",
         },
         "headers": Headers {
           "x-stack-known-error": "USER_EMAIL_ALREADY_EXISTS",
@@ -1172,7 +1174,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": false,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1209,7 +1211,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1261,7 +1263,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1295,7 +1297,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": false,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": false,
           "primary_email_verified": false,
           "profile_image_url": null,
@@ -1353,7 +1355,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1383,7 +1385,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1422,7 +1424,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1461,7 +1463,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1549,7 +1551,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "<stripped UUID>@stack-generated.example.com",
+          "primary_email": "default-mailbox--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1577,13 +1579,13 @@ describe("with server access", () => {
   });
 
   it("should be able to update primary email", async ({ expect }) => {
-    await Project.createAndSwitch({ config: { magic_link_enabled: true }});
     await Auth.Otp.signIn();
+    const mailbox = createMailbox();
     const response = await niceBackendFetch("/api/v1/users/me", {
       accessType: "server",
       method: "PATCH",
       body: {
-        primary_email: "new-primary-email@example.com",
+        primary_email: mailbox.emailAddress,
       },
     });
     expect(response).toMatchInlineSnapshot(`
@@ -1600,7 +1602,7 @@ describe("with server access", () => {
           "oauth_providers": [],
           "otp_auth_enabled": true,
           "passkey_auth_enabled": false,
-          "primary_email": "new-primary-email@example.com",
+          "primary_email": "mailbox-1--<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -1616,24 +1618,125 @@ describe("with server access", () => {
   });
 
   it("should be able to update primary email and sign-in with the new email", async ({ expect }) => {
-    await Project.createAndSwitch();
+    await Auth.Password.signUpWithEmail({ password: "password123" });
+    const mailbox = createMailbox();
+    const response = await niceBackendFetch("/api/v1/users/me", {
+      accessType: "server",
+      method: "PATCH",
+      body: {
+        primary_email: mailbox.emailAddress,
+      },
+    });
+    expect(response.body.primary_email).toEqual(mailbox.emailAddress);
+
+    backendContext.set({
+      mailbox,
+    });
+    await Auth.Password.signInWithEmail({ password: "password123" });
+    expect(response.body.primary_email).toEqual(mailbox.emailAddress);
+  });
+
+  it("should not be able to update primary email to an email already in use for auth", async ({ expect }) => {
+    await Auth.Otp.signIn();
+    const primaryEmail = backendContext.value.mailbox.emailAddress;
+    await Auth.signOut();
+    await bumpEmailAddress();
     await Auth.Password.signUpWithEmail({ password: "password123" });
     const response = await niceBackendFetch("/api/v1/users/me", {
       accessType: "server",
       method: "PATCH",
       body: {
-        primary_email: "new-primary-email@example.com",
+        primary_email: primaryEmail,
       },
     });
-    expect(response.body.primary_email).toEqual("new-primary-email@example.com");
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 400,
+        "body": {
+          "code": "USER_EMAIL_ALREADY_EXISTS",
+          "error": "User email already exists.",
+        },
+        "headers": Headers {
+          "x-stack-known-error": "USER_EMAIL_ALREADY_EXISTS",
+          <some fields may have been hidden>,
+        },
+      }
+    `);
+  });
 
-    backendContext.set({
-      mailbox: {
-        ...backendContext.value.mailbox,
-        emailAddress: "new-primary-email@example.com",
+  it("should not be able to set profile image url to empty string", async ({ expect }) => {
+    await Auth.Otp.signIn();
+    const response = await niceBackendFetch("/api/v1/users/me", {
+      accessType: "server",
+      method: "PATCH",
+      body: {
+        profile_image_url: "",
       },
     });
-    await Auth.Password.signInWithEmail({ password: "password123" });
-    expect(response.body.primary_email).toEqual("new-primary-email@example.com");
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 400,
+        "body": {
+          "code": "SCHEMA_ERROR",
+          "details": { "message": "Request validation failed on PATCH /api/v1/users/me:\\n  - body.profile_image_url is not a valid URL" },
+          "error": "Request validation failed on PATCH /api/v1/users/me:\\n  - body.profile_image_url is not a valid URL",
+        },
+        "headers": Headers {
+          "x-stack-known-error": "SCHEMA_ERROR",
+          <some fields may have been hidden>,
+        },
+      }
+    `);
+  });
+
+  it("should not be able to sign up with an email already in use for auth", async ({ expect }) => {
+    await Auth.Password.signUpWithEmail({ password: "password123" });
+    const response = await niceBackendFetch("/api/v1/auth/password/sign-up", {
+      method: "POST",
+      accessType: "client",
+      body: {
+        email: backendContext.value.mailbox.emailAddress,
+        password: "password123",
+        verification_callback_url: "http://localhost:12345/some-callback-url",
+      },
+    });
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 400,
+        "body": {
+          "code": "USER_EMAIL_ALREADY_EXISTS",
+          "error": "User email already exists.",
+        },
+        "headers": Headers {
+          "x-stack-known-error": "USER_EMAIL_ALREADY_EXISTS",
+          <some fields may have been hidden>,
+        },
+      }
+    `);
+  });
+
+  it("should be able to sign up with an email already in use for auth in a different project", async ({ expect }) => {
+    await Auth.Password.signUpWithEmail({ password: "password123" });
+    await Project.createAndSwitch({});
+    const response = await niceBackendFetch("/api/v1/auth/password/sign-up", {
+      method: "POST",
+      accessType: "client",
+      body: {
+        email: backendContext.value.mailbox.emailAddress,
+        password: "password123",
+        verification_callback_url: "http://localhost:12345/some-callback-url",
+      },
+    });
+    expect(response).toMatchInlineSnapshot(`
+      NiceResponse {
+        "status": 200,
+        "body": {
+          "access_token": <stripped field 'access_token'>,
+          "refresh_token": <stripped field 'refresh_token'>,
+          "user_id": "<stripped UUID>",
+        },
+        "headers": Headers { <some fields may have been hidden> },
+      }
+    `);
   });
 });