+++ packages/stack-shared/src/utils/stores.tsx
@@ -1,6 +1,7 @@
+import { ReadWriteLock } from "./locks";
+import { ReactPromise, pending, rejected, resolved } from "./promises";
 import { AsyncResult, Result } from "./results";
 import { generateUuid } from "./uuids";
-import { ReactPromise, pending, rejected, resolved } from "./promises";
 
 export type ReadonlyStore<T> = {
   get(): T,
@@ -61,6 +62,9 @@ export class Store<T> implements ReadonlyStore<T> {
     return { unsubscribe };
   }
 }
+
+export const storeLock = new ReadWriteLock();
+
 
 export class AsyncStore<T> implements ReadonlyAsyncStore<T> {
   private _isAvailable: boolean;
@@ -171,9 +175,11 @@ export class AsyncStore<T> implements ReadonlyAsyncStore<T> {
   }
 
   async setAsync(promise: Promise<T>): Promise<boolean> {
-    const curCounter = ++this._updateCounter;
-    const result = await Result.fromPromise(promise);
-    return this._setIfLatest(result, curCounter);
+    return await storeLock.withReadLock(async () => {
+      const curCounter = ++this._updateCounter;
+      const result = await Result.fromPromise(promise);
+      return this._setIfLatest(result, curCounter);
+    });
   }
 
   setUnavailable(): void {