+++ packages/stack/src/lib/stack-app.ts
@@ -22,11 +22,11 @@ import { StackAssertionError, concatStacktraces, throwErr } from "@stackframe/st
 import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
 import { DependenciesMap } from "@stackframe/stack-shared/dist/utils/maps";
 import { ProviderType } from "@stackframe/stack-shared/dist/utils/oauth";
-import { deepPlainEquals, filterUndefined, omit } from "@stackframe/stack-shared/dist/utils/objects";
+import { deepPlainEquals, filterUndefined, omit, pick } from "@stackframe/stack-shared/dist/utils/objects";
 import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
 import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
 import { Result } from "@stackframe/stack-shared/dist/utils/results";
-import { Store } from "@stackframe/stack-shared/dist/utils/stores";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
 import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
 import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
 import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
@@ -36,7 +36,6 @@ import React, { useCallback, useMemo } from "react";
 import { constructRedirectUrl } from "../utils/url";
 import { addNewOAuthProviderOrScope, callOAuthCallback, signInWithOAuth } from "./auth";
 import { CookieHelper, createBrowserCookieHelper, createCookieHelper, deleteCookieClient, getCookieClient, setOrDeleteCookie, setOrDeleteCookieClient } from "./cookie";
-
 // NextNavigation.useRouter does not exist in react-server environments and some bundlers try to be helpful and throw a warning. Ignore the warning.
 const NextNavigation = scrambleDuringCompileTime(NextNavigationUnscrambled);
 
@@ -251,6 +250,7 @@ function useStore<T>(store: Store<T>): T {
   return React.useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
 }
 
+/** @internal */
 export const stackAppInternalsSymbol = Symbol.for("StackAuth--DO-NOT-USE-OR-YOU-WILL-BE-FIRED--StackAppInternals");
 
 const allClientApps = new Map<string, [checkString: string, app: StackClientApp<any, any>]>();
@@ -891,7 +891,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
           return Result.error(new KnownErrors.PasskeyRegistrationFailed("Failed to get initiation options for passkey registration"));
         }
 
-        const {options_json, code} = initiationResult.data;
+        const { options_json, code } = initiationResult.data;
 
         // HACK: Override the rpID to be the actual domain
         if (options_json.rp.id !== "THIS_VALUE_WILL_BE_REPLACED.example.com") {
@@ -1474,7 +1474,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
           return Result.error(new KnownErrors.PasskeyAuthenticationFailed("Failed to get initiation options for passkey authentication"));
         }
 
-        const {options_json, code} = initiationResult.data;
+        const { options_json, code } = initiationResult.data;
 
         // HACK: Override the rpID to be the actual domain
         if (options_json.rpId !== "THIS_VALUE_WILL_BE_REPLACED.example.com") {
@@ -1536,12 +1536,14 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
   }
 
   protected async _signOut(session: InternalSession, options?: { redirectUrl?: URL | string }): Promise<void> {
-    await this._interface.signOut(session);
-    if (options?.redirectUrl) {
-      await _redirectTo(options.redirectUrl);
-    } else {
-      await this.redirectToAfterSignOut();
-    }
+    await storeLock.withWriteLock(async () => {
+      await this._interface.signOut(session);
+      if (options?.redirectUrl) {
+        await _redirectTo(options.redirectUrl);
+      } else {
+        await this.redirectToAfterSignOut();
+      }
+    });
   }
 
   async signOut(options?: { redirectUrl?: URL | string }): Promise<void> {
@@ -2266,6 +2268,9 @@ class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string
   private readonly _svixTokenCache = createCache(async () => {
     return await this._interface.getSvixToken();
   });
+  private readonly _metricsCache = createCache(async () => {
+    return await this._interface.getMetrics();
+  });
 
   constructor(options: StackAdminAppConstructorOptions<HasTokenStore, ProjectId>) {
     super({
@@ -2517,6 +2522,35 @@ class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string
 
   protected async _refreshApiKeys() {
     await this._apiKeysCache.refresh([]);
+  }
+
+  get [stackAppInternalsSymbol]() {
+    return {
+      ...super[stackAppInternalsSymbol],
+      useMetrics: (): any => {
+        return useAsyncCache(this._metricsCache, [], "useMetrics()");
+      }
+    };
+  }
+
+  async sendTestEmail(options: {
+    recipientEmail: string,
+    emailConfig: EmailConfig,
+  }): Promise<Result<undefined, { errorMessage: string }>> {
+    const response = await this._interface.sendTestEmail({
+      recipient_email: options.recipientEmail,
+      email_config: {
+        ...(pick(options.emailConfig, ['host', 'port', 'username', 'password'])),
+        sender_email: options.emailConfig.senderEmail,
+        sender_name: options.emailConfig.senderName,
+      },
+    });
+
+    if (response.success) {
+      return Result.ok(undefined);
+    } else {
+      return Result.error({ errorMessage: response.error_message ?? throwErr("Email test error not specified") });
+    }
   }
 }
 
@@ -3410,6 +3444,11 @@ export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId ext
     deleteTeamPermissionDefinition(permissionId: string): Promise<void>,
 
     useSvixToken(): string,
+
+    sendTestEmail(options: {
+      recipientEmail: string,
+      emailConfig: EmailConfig,
+    }): Promise<Result<undefined, { errorMessage: string }>>,
   }
   & StackServerApp<HasTokenStore, ProjectId>
 );
@@ -3453,8 +3492,11 @@ type AsyncStoreProperty<Name extends string, Args extends any[], Value, IsMultip
   & { [key in `${IsMultiple extends true ? "list" : "get"}${Capitalize<Name>}`]: (...args: Args) => Promise<Value> }
   & { [key in `use${Capitalize<Name>}`]: (...args: Args) => Value }
 
-type Prettify<T> = {
-  [K in keyof T]: T[K];
-} & {};
-
-type x = Prettify<CurrentUser>
+type EmailConfig = {
+  host: string,
+  port: number,
+  username: string,
+  password: string,
+  senderEmail: string,
+  senderName: string,
+}