+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @stackframe/stack-backend
 
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-emails@2.7.1
+  - @stackframe/stack-shared@2.7.1
+
+## 2.7.0
+
+### Minor Changes
+
+- Various changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+  - @stackframe/stack-emails@2.7.0
+
+## 2.6.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.39
+  - @stackframe/stack-emails@2.6.39
+
+## 2.6.38
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.38
+  - @stackframe/stack-emails@2.6.38
+
 ## 2.6.37
 
 ### Patch Changes