+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @stackframe/oauth-mock-server
 
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+
+## 2.6.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.39
+
+## 2.6.38
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.38
+
 ## 2.6.37
 
 ### Patch Changes