+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @stackframe/example-middleware-demo
 
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.7.0
+
+## 2.6.39
+
+### Patch Changes
+
+- @stackframe/stack@2.6.39
+
+## 2.6.38
+
+### Patch Changes
+
+- @stackframe/stack@2.6.38
+
 ## 2.6.37
 
 ### Patch Changes