+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @stackframe/stack-emails
 
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.1
+  - @stackframe/stack-ui@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+  - @stackframe/stack-ui@2.7.0
+
+## 2.6.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.39
+  - @stackframe/stack-ui@2.6.39
+
+## 2.6.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.38
+  - @stackframe/stack-ui@2.6.38
+
 ## 2.6.37
 
 ### Patch Changes